Revert "feat: remove inter-dependencies (#44)"
This reverts commit 02ef8afaacaef69233f0da52dc957eae883c0eb6.
