+++ cmd/lambda/main.go
@@ -1,8 +1,14 @@
 package main
 
 import (
+	"os"
+
 	"github.com/aws/aws-lambda-go/lambda"
-	defcon "github.com/kjansson/defcon"
+
+	"github.com/kjansson/defcon"
+	"github.com/kjansson/yac-p/v2/pkg/controller"
+	"github.com/kjansson/yac-p/v2/pkg/prom"
+	"github.com/kjansson/yac-p/v2/pkg/types"
 )
 
 func main() {
@@ -11,14 +17,24 @@ func main() {
 
 func HandleRequest() {
 
-	config := Config{}
-	err := defcon.CheckConfigStruct(&config) // Validate the config struct
+	config := &types.Config{
+		ConfigFileLoader: GetS3Loader(), // Use the S3 loader for Lambda implementation
+		RemoteWriteURL:   os.Getenv("PROMETHEUS_REMOTE_WRITE_URL"),
+		AuthType:         os.Getenv("AUTH_TYPE"),
+		AuthToken:        os.Getenv("TOKEN"),
+		Username:         os.Getenv("USERNAME"),
+		Password:         os.Getenv("PASSWORD"),
+		Region:           os.Getenv("AWS_REGION"),
+		PrometheusRegion: os.Getenv("PROMETHEUS_REGION"),
+		AWSRoleARN:       os.Getenv("AWS_ROLE_ARN"),
+	}
+
+	err := defcon.CheckConfigStruct(config) // Check the config struct for required fields
 	if err != nil {
 		panic(err)
 	}
-	config.ConfigFileLoader = GetS3Loader() // Set the config file loader to S3 for Lambda
 
-	c, err := NewController(config) // Create a new controller instance
+	c, err := controller.NewController(*config) // Create a new controller instance
 	if err != nil {
 		panic(err)
 	}
@@ -41,7 +57,7 @@ func HandleRequest() {
 
 	c.Logger.Log("debug", "Processing metrics")
 	// Process the metrics into timeseries format
-	timeSeries, err := c.ConvertMetrics(metrics)
+	timeSeries, err := prom.ProcessMetrics(metrics, c.Logger)
 	if err != nil {
 		panic(err)
 	}
